+++ server/index.js
@@ -1,4 +1,4 @@
 export {render} from './src/Serialize.js'
-export {memo, lazy, Suspense, Boundary, Fragment, Children} from '../index.js'
-export {createElement, createPortal, cloneElement, isValidElement, h} from '../index.js'
+export {memo, lazy, Suspense, Boundary, Context, Portal, Fragment, Children} from '../index.js'
+export {createElement, cloneElement, isValidElement, h} from '../index.js'
 export {useRef, useMemo, useCallback, useState, useReducer, useContext, useResource, useEffect, useLayout}  from '../index.js'